feat: cameras may now be rendered
Two actions have been added for adding and removing cameras. The dictionary containing the cameras will probably be changed.

Four marker icons have been added to the project. We will probably want to update these in the future.

There is currently a temporary camera manager in the Map component. This manager allows for cameras to be added and removed from the map.

Mocking of Camera class is yet to be added for the "map" unit tests.