fix: bugfix error create project with existing name
This fix changes the code in create_project from returning a None, to raising a valueerror. The related code in project_manager and the test which concerns this function failing have also been changed. A call to create_project which seems to do nothing was removed from test_file_manager. Unsure if serves a purpose, please check.